+++ src/lib/Default/AbstractSmrCombatWeapon.class.php
@@ -6,19 +6,7 @@ abstract class AbstractSmrCombatWeapon {
 	 */
 	const PLANET_DAMAGE_MOD = 0.2;
 
-	protected string $name;
-	protected int $shieldDamage;
-	protected int $armourDamage;
-	protected int $accuracy;
 	protected bool $damageRollover;
-
-	public function getBaseAccuracy() : int {
-		return $this->accuracy;
-	}
-
-	public function getName() : string {
-		return $this->name;
-	}
 
 	/**
 	 * Return the max weapon damage possible in a single round.
@@ -27,13 +15,10 @@ abstract class AbstractSmrCombatWeapon {
 		return max($this->getShieldDamage(), $this->getArmourDamage());
 	}
 
-	public function getShieldDamage() : int {
-		return $this->shieldDamage;
-	}
-
-	public function getArmourDamage() : int {
-		return $this->armourDamage;
-	}
+	abstract public function getBaseAccuracy() : int;
+	abstract public function getName() : string;
+	abstract public function getShieldDamage() : int;
+	abstract public function getArmourDamage() : int;
 
 	public function isDamageRollover() : bool {
 		return $this->damageRollover;
